# [BEAM-10670] Make Spark by default execute Read.Bounded using SplittableDoFn.