--- a/runners/spark/src/main/java/org/apache/beam/runners/spark/SparkRunnerDebugger.java
+++ b/runners/spark/src/main/java/org/apache/beam/runners/spark/SparkRunnerDebugger.java
@@ -26,6 +26,7 @@ import org.apache.beam.runners.spark.translation.TransformTranslator;
 import org.apache.beam.runners.spark.translation.streaming.StreamingTransformTranslator;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.PipelineRunner;
+import org.apache.beam.sdk.options.ExperimentalOptions;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsValidator;
 import org.apache.spark.api.java.JavaSparkContext;
@@ -73,7 +74,16 @@ public final class SparkRunnerDebugger extends PipelineRunner<SparkPipelineResul
 
   @Override
   public SparkPipelineResult run(Pipeline pipeline) {
-    SplittableParDo.convertReadBasedSplittableDoFnsToPrimitiveReads(pipeline);
+    boolean isStreaming =
+        options.isStreaming() || options.as(TestSparkPipelineOptions.class).isForceStreaming();
+    // Default to using the primitive versions of Read.Bounded and Read.Unbounded if we are
+    // executing an unbounded pipeline or the user specifically requested it.
+    if (isStreaming
+        || ExperimentalOptions.hasExperiment(
+            pipeline.getOptions(), "beam_fn_api_use_deprecated_read")
+        || ExperimentalOptions.hasExperiment(pipeline.getOptions(), "use_deprecated_read")) {
+      SplittableParDo.convertReadBasedSplittableDoFnsToPrimitiveReads(pipeline);
+    }
     JavaSparkContext jsc = new JavaSparkContext("local[1]", "Debug_Pipeline");
     JavaStreamingContext jssc =
         new JavaStreamingContext(jsc, new org.apache.spark.streaming.Duration(1000));
@@ -83,9 +93,7 @@ public final class SparkRunnerDebugger extends PipelineRunner<SparkPipelineResul
     TransformTranslator.Translator translator = new TransformTranslator.Translator();
 
     SparkNativePipelineVisitor visitor;
-    if (options.isStreaming()
-        || (options instanceof TestSparkPipelineOptions
-            && ((TestSparkPipelineOptions) options).isForceStreaming())) {
+    if (isStreaming) {
       SparkPipelineTranslator streamingTranslator =
           new StreamingTransformTranslator.Translator(translator);
       EvaluationContext ctxt = new EvaluationContext(jsc, pipeline, options, jssc);