@lukecwik 2020-10-06
R: @iemejia 

Part 2 of https://github.com/apache/beam/pull/12603

Note that experimenting with watermark holds will likely take a few days to see if some of my ideas will work.