@lukecwik 2020-10-06
Run Spark StructuredStreaming ValidatesRunner