@lukecwik 2020-10-06
Run Spark Runner Nexmark Tests